WIP. Refactorized and prepared code to have reservation link
