+++ commands/cancel_command.py
@@ -1,0 +1,22 @@
+import logging
+from telegram import ReplyKeyboardMarkup, ReplyKeyboardRemove, Update
+from telegram.ext import (
+    Application,
+    CommandHandler,
+    ContextTypes,
+    ConversationHandler,
+    MessageHandler,
+    filters,
+)
+
+logger = logging.getLogger(__name__)
+
+async def cancel(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+    """Cancels and ends the conversation."""
+    user = update.message.from_user
+    logger.info("User %s canceled the conversation.", user.first_name)
+    await update.message.reply_text(
+        "Conversación finalizada.", reply_markup=ReplyKeyboardRemove()
+    )
+
+    return ConversationHandler.END
