+++ main.py
@@ -1,9 +1,6 @@
-import datetime
 import json
 import logging
 from typing import Final
-import httpx
-from datetime import datetime, timedelta
 
 from telegram import ReplyKeyboardMarkup, ReplyKeyboardRemove, Update
 from telegram.ext import (
@@ -14,6 +11,9 @@ from telegram.ext import (
     MessageHandler,
     filters,
 )
+
+from commands.free_slots_command import *
+from commands.cancel_command import *
 
 # Enable logging
 logging.basicConfig(
@@ -34,86 +34,6 @@ TOKEN: Final = secrets["TELEGRAM_BOT_TOKEN"]
 del secrets
 
 #endregion
-
-
-
-
-CHECKAVAILABILITY = range(1)
-
-
-async def start(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    user = update.message.from_user
-    logger.info(f"{user.first_name} with id {user.id} started a conversation")
-
-    reply_keyboard = [["27/02", "28/02", "29/02"]]
-
-    await update.message.reply_text(
-        "Qué hace mostro? Sacate el plug de la cola y prestame atención. "
-        "Podes escribir /cancel si te hinchaste los huevos.\n\n"
-        "Para qué fecha (dd/mm) queres que te diga los turnos disponibles?",
-        reply_markup=ReplyKeyboardMarkup(
-            reply_keyboard, one_time_keyboard=True, input_field_placeholder="dd/mm"
-        ),
-    )
-
-    return CHECKAVAILABILITY
-
-
-async def check_availability(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    user = update.message.from_user
-    try:
-        date_string = f"{update.message.text}/{datetime.today().year}"
-        date_object = datetime.strptime(date_string, "%d/%m/%Y")  
-    except:
-        await update.message.reply_text("Te crees piola? Ingresá bien la fecha, payaso.", reply_markup=ReplyKeyboardRemove())
-        return CHECKAVAILABILITY
-
-    formatted_date = date_object.strftime("%Y-%m-%d")
-    url = f"https://alquilatucancha.com/api/v3/availability/sportclubs/678?date={formatted_date}"
-
-    async with httpx.AsyncClient() as client:
-        response = await client.get(url)
-        if response.status_code == 200:
-            availability_data = response.json()
-            message_list = retrieve_free_slots(availability_data)
-            message = "\n".join(message_list)
-            logger.info(f"{user.first_name} retrieved the info correctly")
-        else:
-            message = "Error fetching availability data."
-            logger.warning(f"{user.first_name} couldn't retrieve the info")
-
-    await update.message.reply_text(f"{message}", reply_markup=ReplyKeyboardRemove())
-    logger.info(f"{user.first_name} finished the conversation")
-    return ConversationHandler.END
-
-def retrieve_free_slots(availability_data):
-    available_courts = availability_data.get("available_courts", [])
-    message_list = ["Available slots:"]
-    previous_slot = ""
-    for court in available_courts:
-        court_name = court.get("name", "Unknown court")
-        message_list.append(f"\n{court_name}:")
-        for slot in court.get("available_slots", []):
-            if slot.get("start", "Unknown start time") == previous_slot:
-                message = message_list.pop()
-                message_list.append(f"{message}, {slot.get('duration', 0)} mins")
-                continue
-            start_time = slot.get("start", "Unknown start time")
-            duration = slot.get("duration", 0)
-            start_time_obj = datetime.fromisoformat(start_time)
-            message_list.append(f"  - Start: {start_time_obj.strftime('%H:%M')}, Duration: {duration} mins")
-            previous_slot = start_time
-    return message_list
-
-async def cancel(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    """Cancels and ends the conversation."""
-    user = update.message.from_user
-    logger.info("User %s canceled the conversation.", user.first_name)
-    await update.message.reply_text(
-        "Conversación finalizada.", reply_markup=ReplyKeyboardRemove()
-    )
-
-    return ConversationHandler.END
 
 
 def main() -> None:
